+++ Project/BankIS/src/main/java/bis/project/controllers/ClientDetailsController.java
@@ -3,20 +3,27 @@ package bis.project.controllers;
 import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import bis.project.model.ClientDetails;
 import bis.project.services.ClientDetailsServices;
+import bis.project.services.CredentialsServices;
 
 @RestController
 public class ClientDetailsController {
 	
 	@Autowired
 	private ClientDetailsServices services;
+	
+	@Autowired
+	private CredentialsServices cServices;
 	
 	@RequestMapping(value = "/api/clients", 
 					method = RequestMethod.GET)
@@ -27,26 +34,53 @@ public class ClientDetailsController {
 	
 	@RequestMapping(value = "/api/clients/{id}", 
 					method = RequestMethod.GET)
-	public ClientDetails getClient(@PathVariable("id") Integer id) {
-		return services.getClient(id);
+	public ResponseEntity<ClientDetails> getClient(@PathVariable("id") Integer id) {
+		ClientDetails  client = services.getClient(id);
+		
+		if(client != null) {
+			return new ResponseEntity<ClientDetails>(client, HttpStatus.OK);
+		}
+		
+		return new ResponseEntity<ClientDetails>(HttpStatus.NOT_FOUND); 
 	}
 	
 	@RequestMapping(value = "/api/clients", 
 					method = RequestMethod.POST)
-	public ClientDetails addClient(@RequestBody ClientDetails client) {
-		return services.addClient(client);
+	public ResponseEntity<ClientDetails> addClient(@RequestBody ClientDetails client, @RequestHeader(value="Authorization") String basicAuth) {
+		boolean isAuthorized = cServices.isAuthorized(basicAuth, "addClient");
+		
+		if(isAuthorized) {
+			ClientDetails c = services.addClient(client);
+			return new ResponseEntity<ClientDetails>(c, HttpStatus.OK);
+		}
+		
+		return new ResponseEntity<ClientDetails>(HttpStatus.UNAUTHORIZED);
 	}
 	
 	@RequestMapping(value = "/api/clients/{id}", 
 					method = RequestMethod.PUT)
-	public ClientDetails updateClient(@PathVariable("id") Integer id, @RequestBody ClientDetails client) {
-		client.setId(id);
-		return services.updateClient(client);
+	public ResponseEntity<ClientDetails> updateClient(@PathVariable("id") Integer id, @RequestBody ClientDetails client, @RequestHeader(value="Authorization") String basicAuth) {
+		boolean isAuthorized = cServices.isAuthorized(basicAuth, "updateClient");
+		
+		if(isAuthorized) {
+			client.setId(id);
+			ClientDetails c = services.updateClient(client);
+			return new ResponseEntity<ClientDetails>(c, HttpStatus.OK);
+		}
+		
+		return new ResponseEntity<ClientDetails>(HttpStatus.UNAUTHORIZED);
 	}
 	
 	@RequestMapping(value = "/api/clients/{id}", 
 					method = RequestMethod.DELETE)
-	public void deleteClient(@PathVariable("id") Integer id) {
-		services.deleteClient(id);
+	public ResponseEntity<ClientDetails> deleteClient(@PathVariable("id") Integer id, @RequestHeader(value="Authorization") String basicAuth) {
+		boolean isAuthorized = cServices.isAuthorized(basicAuth, "deleteClient");
+		
+		if(isAuthorized) {
+			services.deleteClient(id);
+			return new ResponseEntity<ClientDetails>(HttpStatus.OK);
+		}
+		
+		return new ResponseEntity<ClientDetails>(HttpStatus.UNAUTHORIZED);
 	}
 }
