+++ Project/BankIS/src/main/java/bis/project/controllers/WorkTypeController.java
@@ -3,13 +3,17 @@ package bis.project.controllers;
 import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import bis.project.model.WorkType;
+import bis.project.services.CredentialsServices;
 import bis.project.services.WorkTypeServices;
 
 @RestController
@@ -17,6 +21,9 @@ public class WorkTypeController {
 	
 	@Autowired
 	private WorkTypeServices services;
+	
+	@Autowired
+	private CredentialsServices cServices;
 	
 	@RequestMapping(value = "/api/worktypes", 
 					method = RequestMethod.GET)
@@ -27,26 +34,53 @@ public class WorkTypeController {
 	
 	@RequestMapping(value = "/api/worktypes/{id}", 
 					method = RequestMethod.GET)
-	public WorkType getWorkType(@PathVariable("id") Integer id) {
-		return services.getWorkType(id);
+	public ResponseEntity<WorkType> getWorkType(@PathVariable("id") Integer id) {
+		WorkType workType = services.getWorkType(id);
+		
+		if(workType != null) {
+			return new ResponseEntity<WorkType>(workType, HttpStatus.OK);
+		}
+		
+		return new ResponseEntity<WorkType>(HttpStatus.NOT_FOUND); 
 	}
 	
 	@RequestMapping(value = "/api/worktypes", 
 					method = RequestMethod.POST)
-	public WorkType addWorkType(@RequestBody WorkType workType) {
-		return services.addWorkType(workType);
+	public ResponseEntity<WorkType> addWorkType(@RequestBody WorkType workType, @RequestHeader(value="Authorization") String basicAuth) {
+		boolean isAuthorized = cServices.isAuthorized(basicAuth, "addWorkType");
+		
+		if(isAuthorized) {
+			WorkType wt = services.addWorkType(workType);
+			return new ResponseEntity<WorkType>(wt, HttpStatus.OK);
+		}
+		
+		return new ResponseEntity<WorkType>(HttpStatus.UNAUTHORIZED); 
 	}
 	
 	@RequestMapping(value = "/api/worktypes/{id}", 
 					method = RequestMethod.PUT)
-	public WorkType updateWorkType(@PathVariable("id") Integer id, @RequestBody WorkType workType) {
-		workType.setId(id);
-		return services.updateWorkType(workType);
+	public ResponseEntity<WorkType> updateWorkType(@PathVariable("id") Integer id, @RequestBody WorkType workType, @RequestHeader(value="Authorization") String basicAuth) {
+		boolean isAuthorized = cServices.isAuthorized(basicAuth, "updateWorkType");
+		
+		if(isAuthorized) {
+			workType.setId(id);
+			WorkType wt = services.updateWorkType(workType);
+			return new ResponseEntity<WorkType>(wt, HttpStatus.OK);
+		}
+		
+		return new ResponseEntity<WorkType>(HttpStatus.UNAUTHORIZED);
 	}
 	
 	@RequestMapping(value = "/api/worktypes/{id}", 
 					method = RequestMethod.DELETE)
-	public void deleteWorkType(@PathVariable("id") Integer id) {
-		services.deleteWorkType(id);
+	public ResponseEntity<WorkType> deleteWorkType(@PathVariable("id") Integer id, @RequestHeader(value="Authorization") String basicAuth) {
+		boolean isAuthorized = cServices.isAuthorized(basicAuth, "deleteWorkType");
+		
+		if(isAuthorized) {
+			services.deleteWorkType(id);
+			return new ResponseEntity<WorkType>(HttpStatus.OK);
+		}
+		
+		return new ResponseEntity<WorkType>(HttpStatus.UNAUTHORIZED);
 	}
 }
